% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_serenity_viz.R
\name{addPlots}
\alias{addPlots}
\title{Add plots with data to plot list}
\usage{
addPlots(plots, primitive = FALSE, data_dim = 0, data_types = "")
}
\arguments{
\item{plots}{List of plots (name is geom, value is name - don't ask)}

\item{primitive}{Is this geom a primitive?}

\item{data_dim}{Number of data dimensions}

\item{data_types}{String of "C"s and "D"s}
}
\value{
data.frame
}
\description{
When recipes are implemented, this will be refactored to include recipe
  information, such as which geoms and arguments (and/or) which plots.
}
