% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bs_accordion}
\alias{bs_accordion}
\title{Wrapper for bsplus::bs_accordion}
\usage{
bs_accordion(id, panel_type = "default", use_heading_link = TRUE, class = NULL)
}
\arguments{
\item{id}{character, unique id for accordion \verb{<div/>},
also serves as root id for panels appended using \code{bs_append()}}

\item{panel_type}{character, one of the standard Bootstrap types
\code{c("default", "primary", "success", "info", "warning", "danger")}}

\item{use_heading_link}{logical, indicates whether to make the entire panel heading
clickable.}

\item{class}{Additional classes to add to accordion}
}
\value{
bs_accordion
}
\description{
Adds appropriate class for rendering
}
