% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{create_aes_input}
\alias{create_aes_input}
\title{Create aes inputs}
\usage{
create_aes_input(inputId, aes, aes_val, default = "")
}
\arguments{
\item{inputId}{Id of Shiny input}

\item{aes}{Name of aesthetic}

\item{aes_val}{Value of aesthetic}

\item{default}{Not used}
}
\description{
Create aes inputs
}
