% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{createStageStateChangeEvent}
\alias{createStageStateChangeEvent}
\title{Creates observeEvent to change state of stage button}
\usage{
createStageStateChangeEvent(
  stage = "start",
  mapping,
  post_aesthetics_render_update,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{stage}{Which stage?}

\item{mapping}{Mapping}

\item{session}{Session}
}
\value{
observeEvent
}
\description{
Creates observeEvent to change state of stage button
}
