% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{dataComputedUI}
\alias{dataComputedUI}
\title{UI for computed data}
\usage{
dataComputedUI(id, stat = "identity")
}
\arguments{
\item{id}{Data ID}

\item{stat}{Currently active stat}
}
\value{
UI for data
}
\description{
UI for computed data
}
