% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{dataServer}
\alias{dataServer}
\title{Server for data module}
\usage{
dataServer(input, output, session, dataset)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{dataset}{Passed in dataset for visualization}
}
\description{
Server for data module
}
