% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_var.R
\name{dataVarServer}
\alias{dataVarServer}
\title{Server for data variable module}
\usage{
dataVarServer(id, var)
}
\arguments{
\item{id}{ID of data variable module}

\item{var}{Variable from data frame}
}
\description{
Server for data variable module
}
