% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_var.R
\name{dataVarUI}
\alias{dataVarUI}
\title{UI for data variable submodule}
\usage{
dataVarUI(id, var, default = "")
}
\arguments{
\item{id}{ID of data variable}

\item{var}{Variable from data frame}

\item{default}{Some kind of class}
}
\value{
UI for data variable
}
\description{
UI for data variable submodule
}
