% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropZone.R
\name{dropZoneInput}
\alias{dropZoneInput}
\title{Create a drop zone for drag-and-drop}
\usage{
dropZoneInput(inputId, ..., class = "dropzone", style = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or `\code{NULL} for no label.}
}
\description{
Create an input control for drag-and-drop behavior
}
