% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getLayerInfo}
\alias{getLayerInfo}
\title{Get layer information from namespace string}
\usage{
getLayerInfo(ns)
}
\arguments{
\item{ns}{Namespace function}
}
\value{
List of layer information, including app name, plot id, geom,
  layer id, and aesthetic
}
\description{
Get layer information from namespace string
}
