% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_labels.R
\name{labelsServer}
\alias{labelsServer}
\title{Server for labels module}
\usage{
labelsServer(input, output, session, xlabel, ylabel)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{xlabel}{x-label reactive to replace placeholder}

\item{ylabel}{y-label reactive to replace placeholder}
}
\value{
Reactive expression for label code (string)
}
\description{
Server for labels module
}
