% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_custom.R
\name{layerAesCustomServer}
\alias{layerAesCustomServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesCustomServer(
  id,
  custom_for,
  custom_value = reactive({     NULL }),
  waitforit = 0,
  aesUpdateDependency = reactive({     NULL })
)
}
\arguments{
\item{id}{ID of layer aesthetic custom submodule}

\item{custom_for}{Reactive character vector to customize}

\item{custom_value}{Reactive character vector to initialize and override}

\item{waitforit}{If custom_for input is within renderUI, set to 1 to make
sure custom_for does not override custom_value.  REFACTOR: This is REALLY
cludgy. Couldn't figure out a more elegant way around this issue.}

\item{aesUpdateDependency}{Trigger update on layer change}
}
\value{

}
\description{
Server for layer aesthetic submodule
}
