% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_custom.R
\name{layerAesCustomUI}
\alias{layerAesCustomUI}
\title{UI for layer aesthetic custom submodule}
\usage{
layerAesCustomUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic custom submodule}
}
\value{
UI for layer aesthetic custom submodule
}
\description{
UI for layer aesthetic custom submodule
}
