% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_mapping_stage.R
\name{layerAesMappingStageServer}
\alias{layerAesMappingStageServer}
\title{Server for layer aesthetic stage submodule}
\usage{
layerAesMappingStageServer(
  id,
  stage,
  choices,
  default = list(),
  inherit = reactive({     TRUE }),
  linked = reactive({     TRUE }),
  aesUpdateDependency = reactive({     NULL })
)
}
\arguments{
\item{id}{ID of layer aesthetic stage submodule}

\item{stage}{Which stage? Possible values include \code{start}, \code{after_stat}, and \code{after_scale}.}

\item{choices}{Reactive: Stage choices (given by \code{dataInputChoices})}

\item{default}{Reactive or list of reactive default values (\code{mapping},
\code{custom_mapping}, and \code{custom_toggle}). Defaults to NULL for all three. If
a single reactive, then understood as specifying mapping only.}

\item{inherit}{Is this layer allowing aesthetic inheritance?}

\item{linked}{Is this aesthetic inheriting from base?}

\item{aesUpdateDependency}{Trigger update on layer change}
}
\value{
List containing layer aesthetic mapping stage module state.
\describe{
  \item{mapping}{Reactive expression of stage mapping state}
  \item{custom_mapping}{Reactive expression of stage custom mapping state (used for initialization of \code{default})}
  \item{custom_toggle}{Reactive expression of stage custom toggle state (used for initialization of \code{default})}
  \item{changed}{Reactive expression triggered when mapping changes (used for unlinking to base layer)}
  \item{code}{Reactive expression of aesthetic mapping stage code (string)}
}
}
\description{
Server for layer aesthetic stage submodule
}
