% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_mapping_stage.R
\name{layerAesMappingStageUI}
\alias{layerAesMappingStageUI}
\title{UI for layer aesthetic stage submodule}
\usage{
layerAesMappingStageUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic stage submodule}
}
\value{
UI for layer aesthetic stage submodule
}
\description{
UI for layer aesthetic stage submodule
}
