% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_mapping.R
\name{layerAesMappingUI}
\alias{layerAesMappingUI}
\title{UI for layer aesthetic mapping submodule}
\usage{
layerAesMappingUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic mapping submodule}
}
\value{
UI for layer aesthetic mapping submodule
}
\description{
UI for layer aesthetic mapping submodule
}
