% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{layerAesServer}
\alias{layerAesServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesServer(input, output, session, aesUpdateDependency,
  geom_blank_input, inherit.aes, default_geom_aes, default_stat_aes,
  dataset, computed_vars)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{aesUpdateDependency}{Trigger update on layer change}

\item{geom_blank_input}{Need geom_blank values to check for inheritance}

\item{inherit.aes}{Reactive: Is this aesthetic inheritable?}

\item{default_geom_aes}{Default value for geom aesthetic}

\item{default_stat_aes}{Reactive value of default stat aesthetic - might be mapping!}

\item{dataset}{Dataset}

\item{computed_vars}{Reactive value of stat computed variables}
}
\description{
Server for layer aesthetic submodule
}
