% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aesthetics.R
\name{layerAestheticsServer}
\alias{layerAestheticsServer}
\title{Server for layer aesthetics module}
\usage{
layerAestheticsServer(
  id,
  layer_id,
  geom,
  layer_selected,
  base_layer_stages,
  dataset,
  inherit_aes,
  layer_stat
)
}
\arguments{
\item{id}{ID of layer aesthetics submodule}

\item{layer_id}{Layer ID}

\item{geom}{Geom}

\item{layer_selected}{Reactive value of ID of currently selected layer}

\item{base_layer_stages}{Need base layer aesthetic mappings for inheritance}

\item{dataset}{Dataset}

\item{inherit_aes}{Reactive: Is this aesthetic inheritable?}

\item{layer_stat}{Reactive value of currently selected layer stat}
}
\description{
Server for layer aesthetics module
}
