% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aesthetics.R
\name{layerAestheticsUI}
\alias{layerAestheticsUI}
\title{UI for layer aesthetics module}
\usage{
layerAestheticsUI(id)
}
\arguments{
\item{id}{Layer ID}
}
\value{
UI for layer
}
\description{
UI for layer aesthetics module
}
