% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_value.R
\name{layerAesValueServer}
\alias{layerAesValueServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesValueServer(
  id,
  aesthetic,
  initial,
  show_initial = reactive({     FALSE }),
  aesUpdateDependency = reactive({     NULL })
)
}
\arguments{
\item{id}{ID of layer aesthetic value submodule}

\item{aesthetic}{Input will be for this aesthetic (determines widget)}

\item{initial}{Input initial value}

\item{show_initial}{Reactive: Show initial value?}

\item{aesUpdateDependency}{Trigger update on layer change}
}
\value{
String "<aesthetic> = <value>"
}
\description{
Server for layer aesthetic submodule
}
