% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_value.R
\name{layerAesValueUI}
\alias{layerAesValueUI}
\title{UI for layer aesthetic value submodule}
\usage{
layerAesValueUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic value submodule}
}
\value{
UI for layer aesthetic value submodule
}
\description{
UI for layer aesthetic value submodule
}
