% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerChoiceUI}
\alias{layerChoiceUI}
\title{UI for layer choice}
\usage{
layerChoiceUI(plot_id)
}
\arguments{
\item{geom}{Layer geom}
}
\value{
UI for layer choice
}
\description{
UI for layer choice
}
