% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_params.R
\name{layerParamsUI}
\alias{layerParamsUI}
\title{UI for layer parameters submodule}
\usage{
layerParamsUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic}
}
\value{
UI for layer parameters
}
\description{
UI for layer parameters submodule
}
