% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_position.R
\name{layerPositionServer}
\alias{layerPositionServer}
\title{Server function for position arguments}
\usage{
layerPositionServer(input, output, session, base_data, default_position)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{base_data}{Pre-isolated reactive to base layer data}

\item{default_position}{Default layer position}
}
\description{
Server function for position arguments
}
