% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layers.R
\name{layersServer}
\alias{layersServer}
\title{Server for all layers module}
\usage{
layersServer(id, dataset)
}
\arguments{
\item{id}{ID of layers module}

\item{dataset}{Dataset}
}
\value{
List containing layers module state.
\describe{
  \item{code}{Reactive expression of layers code (string)}
  \item{selected_layer}{Reactive expression of currently selected layer (string)}
  \item{selected_stat}{Reactive expression of currently selected layer (string)}
  \item{aesthetics}{Reactive expression of selected layer aesthetics (vector of strings)}
}
}
\description{
Server for all layers module
}
