% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layers.R
\name{layersUI}
\alias{layersUI}
\title{Module UI for all layers}
\usage{
layersUI(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Layers tab ID}

\item{session}{Shiny user session}
}
\value{
UI for layer
}
\description{
Module UI for all layers
}
