% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_serenity_viz.R
\name{serenityVizServer}
\alias{serenityVizServer}
\title{Server code for Serenity Viz module}
\usage{
serenityVizServer(id, dataset)
}
\arguments{
\item{id}{ID of Serenity Viz module}

\item{dataset}{Passed in dataset for visualization}
}
\description{
Server code for Serenity Viz module
}
