% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_serenity_viz.R
\name{serenityVizUI}
\alias{serenityVizUI}
\title{UI for Serenity Viz module}
\usage{
serenityVizUI(id, dataset)
}
\arguments{
\item{id}{Serenity Viz module ID}

\item{dataset}{Passed in dataset for visualization}
}
\value{
UI for Serenity Viz module
}
\description{
UI for Serenity Viz module
}
