% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_value.R
\name{update_aes_input}
\alias{update_aes_input}
\title{Update aes inputs}
\usage{
update_aes_input(session, inputId, aes, aes_val)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{Id of input object}

\item{aes}{Name of aesthetic}

\item{aes_val}{Value of aesthetic}
}
\description{
Update aes inputs
}
