% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_value.R
\name{create_aes_input}
\alias{create_aes_input}
\title{Create aes inputs}
\usage{
create_aes_input(inputId, aes, aes_val, class = NULL)
}
\arguments{
\item{inputId}{Id of Shiny input}

\item{aes}{Name of aesthetic}

\item{aes_val}{Value of aesthetic}

\item{class}{Class to add to aesthetic wrapper div}
}
\description{
Create aes inputs
}
