% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{dataInputChoices}
\alias{dataInputChoices}
\title{UI for data choices, either for dropzone or picker}
\usage{
dataInputChoices(vars = NULL, zone = "varzone",
  session = getDefaultReactiveDomain())
}
\arguments{
\item{vars}{Dataset, if not computed; otherwise, element of stat_computed_vars}

\item{zone}{"varzone" or "aeszone"}

\item{session}{Shiny user session}
}
\description{
UI for data choices, either for dropzone or picker
}
