% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{dataServer}
\alias{dataServer}
\title{Server for data module}
\usage{
dataServer(id, dataset, layers)
}
\arguments{
\item{id}{ID of data module}

\item{dataset}{Passed in dataset for visualization}

\item{layers}{State object of layers module}
}
\description{
Server for data module
}
