% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_var.R
\name{dataVarServer}
\alias{dataVarServer}
\title{Server for data variable submodule}
\usage{
dataVarServer(input, output, session, var)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{var}{Variable from data frame}
}
\description{
Server for data variable submodule
}
