% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_mapping.R
\name{layerAesMappingServer}
\alias{layerAesMappingServer}
\title{Server for layer aesthetic stage submodule}
\usage{
layerAesMappingServer(id, stage, geom, aesthetics, computed_vars)
}
\arguments{
\item{id}{ID of layer aesthetic stage submodule}

\item{stage}{Stage}

\item{geom}{Geom}

\item{aesthetics}{Aesthetics}

\item{computed_vars}{Reactive value of stat computed variables}
}
\value{

}
\description{
Server for layer aesthetic stage submodule
}
