% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes_mapping_stage.R
\name{layerAesMappingStageServer}
\alias{layerAesMappingStageServer}
\title{Server for layer aesthetic stage submodule}
\usage{
layerAesMappingStageServer(
  id,
  stage,
  choices,
  default = list(),
  inherit = reactive({     TRUE }),
  linked = reactive({     TRUE })
)
}
\arguments{
\item{id}{ID of layer aesthetic stage submodule}

\item{stage}{Which stage?}

\item{choices}{Reactive: Stage choices (given by dataInputChoices)}

\item{default}{Reactive or list of reactive default values (mapping,
custom_mapping, and custom_toggle). Defaults to NULL for all three. If
a single reactive, then understood as specifying mapping only.}

\item{inherit}{Is this layer allowing aesthetic inheritance?}

\item{linked}{Is this aesthetic inheriting from base?}
}
\value{

}
\description{
Server for layer aesthetic stage submodule
}
