% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{layerAesServer}
\alias{layerAesServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesServer(
  id,
  geom,
  aesthetic,
  base_layer_stages,
  inherit_aes,
  default_geom_aes,
  default_stat_aes,
  required,
  dataset,
  computed_vars,
  aesthetics
)
}
\arguments{
\item{id}{ID of layer aesthetic submodule}

\item{geom}{Geom}

\item{aesthetic}{Aesthetic}

\item{base_layer_stages}{Need base layer aesthetic mapping to use for inheritance}

\item{inherit_aes}{Reactive: Is this aesthetic inheritable?}

\item{default_geom_aes}{Default value for geom aesthetic}

\item{default_stat_aes}{Reactive value of default stat aesthetic - might be mapping!}

\item{required}{Reactive value of required status for aesthetic}

\item{dataset}{Dataset}

\item{computed_vars}{Reactive value of stat computed variables}

\item{aesthetics}{Reactive value of aesthetics (combines layer and stat aesthetics)}
}
\description{
Server for layer aesthetic submodule
}
