% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{layerAesServer}
\alias{layerAesServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesServer(input, output, session, aesUpdateDependency,
  base_layer_mapping, inherit.aes, default_geom_aes, default_stat_aes,
  required, dataset, computed_vars)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{aesUpdateDependency}{Trigger update on layer change}

\item{base_layer_mapping}{Need base layer aesthetic mapping to use for inheritance}

\item{inherit.aes}{Reactive: Is this aesthetic inheritable?}

\item{default_geom_aes}{Default value for geom aesthetic}

\item{default_stat_aes}{Reactive value of default stat aesthetic - might be mapping!}

\item{required}{Reactive value of required status for aesthetic}

\item{dataset}{Dataset}

\item{computed_vars}{Reactive value of stat computed variables}
}
\description{
Server for layer aesthetic submodule
}
