% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{layerAesServer}
\alias{layerAesServer}
\title{Server for layer aesthetic submodule}
\usage{
layerAesServer(input, output, session, triggerAesUpdate, geom_blank_input,
  inherit.aes, default_aes, dataset, renderNum)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{triggerAesUpdate}{Trigger update on layer change}

\item{geom_blank_input}{Need geom_blank values to check for inheritance}

\item{inherit.aes}{Reactive: Is this aesthetic inheritable?}

\item{default_aes}{Default value for aesthetic}

\item{dataset}{Dataset}

\item{renderNum}{Closure used to mark number of times aesthetic is rendered}
}
\description{
Server for layer aesthetic submodule
}
