% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aesthetics.R
\name{layerAestheticsServer}
\alias{layerAestheticsServer}
\title{Server for layer aesthetics module}
\usage{
layerAestheticsServer(input, output, session, layers_selected,
  geom_blank_input, dataset)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{layers_selected}{Reactive value of currently selected layer}

\item{geom_blank_input}{Need geom_blank values to check for inheritance}

\item{dataset}{Dataset}
}
\description{
Server for layer aesthetics module
}
