% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aesthetics.R
\name{layerAestheticsServer}
\alias{layerAestheticsServer}
\title{Server for layer aesthetics module}
\usage{
layerAestheticsServer(
  id,
  layer_id,
  geom,
  selected_layer,
  base_layer_aesthetics,
  dataset,
  inherit_aes,
  layer_stat,
  post_aesthetics_render_update
)
}
\arguments{
\item{id}{ID of layer aesthetics submodule}

\item{layer_id}{Layer ID}

\item{geom}{Geom}

\item{selected_layer}{Reactive value of ID of currently selected layer}

\item{base_layer_aesthetics}{Need base layer aesthetic mappings for inheritance}

\item{dataset}{Dataset}

\item{inherit_aes}{Reactive: Is this aesthetic inheritable?}

\item{layer_stat}{Reactive value of currently selected layer stat}

\item{post_aesthetics_render_update}{JS->R reactive to signal post-aesthetic render update (tracks state in UI)}
}
\value{
List containing layer aesthetics module state.
\describe{
  \item{code}{Reactive expression of layer aesthetics code (string)}
  \item{aesthetics}{Reactive expression of selected layer aesthetics (vector of strings)}
}
}
\description{
Server for layer aesthetics module
}
