% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aesthetics.R
\name{layerAestheticsServer}
\alias{layerAestheticsServer}
\title{Server for layer aesthetics module}
\usage{
layerAestheticsServer(input, output, session, layer_selected,
  base_layer_mappings, dataset, inherit.aes, layer_stat)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{layer_selected}{Reactive value of currently selected layer}

\item{base_layer_mappings}{Need base layer aesthetic mappings for inheritance}

\item{dataset}{Dataset}

\item{inherit.aes}{Inherit aesthetics from base layer? (reactive)}

\item{layer_stat}{Reactive value of currently selected layer stat}
}
\description{
Server for layer aesthetics module
}
