% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_aes.R
\name{layerAesUI}
\alias{layerAesUI}
\title{UI for layer aesthetic submodule}
\usage{
layerAesUI(id)
}
\arguments{
\item{id}{ID of layer aesthetic}
}
\value{
UI for layer aesthetic
}
\description{
UI for layer aesthetic submodule
}
