% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_params.R
\name{layerParamsServer}
\alias{layerParamsServer}
\title{Server for layer parameters submodule}
\usage{
layerParamsServer(id, base_data, layer_stat)
}
\arguments{
\item{id}{ID of layer parameter module}

\item{base_data}{Reactive values of ggplot plot object states (data and scales)}

\item{layer_stat}{Reactive value of currently selected layer stat}
}
\value{
List containing layer parameter module state.
\describe{
  \item{inherit.aes}{Reactive expression of aesthetic inheritance layer setting (boolean)}
  \item{code}{Reactive expression of layer parameter code}
}
}
\description{
Server for layer parameters submodule
}
