% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_params.R
\name{layerParamsServer}
\alias{layerParamsServer}
\title{Server for layer parameters submodule}
\usage{
layerParamsServer(input, output, session, base_data)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{base_data}{Reactive values of ggplot plot object states (data and scales)}
}
\description{
Server for layer parameters submodule
}
