% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer_position.R
\name{layerPositionServer}
\alias{layerPositionServer}
\title{Server function for position arguments}
\usage{
layerPositionServer(id, base_data, default_position)
}
\arguments{
\item{id}{ID of layer position module}

\item{base_data}{Pre-isolated reactive to base layer data}

\item{default_position}{Default layer position}
}
\value{
Reactive expression for layer position code (string)
}
\description{
Server function for position arguments
}
