% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerServer}
\alias{layerServer}
\title{Server for layers module}
\usage{
layerServer(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}
}
\description{
Server for layers module
}
