% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerServer}
\alias{layerServer}
\title{Server for layer module}
\usage{
layerServer(
  id,
  selected_layer,
  base_layer_aesthetics,
  dataset,
  post_aesthetics_render_update,
  ggbase = NULL
)
}
\arguments{
\item{id}{ID of layer module}

\item{selected_layer}{Reactive value of currently selected layer}

\item{base_layer_aesthetics}{Need base layer aesthetic values to check for inheritance}

\item{dataset}{Dataset}

\item{post_aesthetics_render_update}{JS->R reactive to signal post-aesthetic render update (tracks state in UI)}

\item{ggbase}{All your base are belong to us...}
}
\value{
List containing individual layer module state.
\describe{
  \item{code}{Reactive expression of layer code (string)}
  \item{stat}{Reactive expression of currently selected stat (string)}
  \item{aesthetics}{Reactive expression of layer aesthetics (vector of strings)}
}
}
\description{
Server for layer module
}
