% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerServer}
\alias{layerServer}
\title{Server for layer module}
\usage{
layerServer(id, layers_selected, geom_blank_input, dataset, ggbase = NULL)
}
\arguments{
\item{id}{ID of layer module}

\item{layers_selected}{Reactive value of currently selected layer}

\item{geom_blank_input}{Need geom_blank values to check for inheritance}

\item{dataset}{Dataset}

\item{ggbase}{All your base are belong to us...}
}
\description{
Server for layer module
}
