% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerServer}
\alias{layerServer}
\title{Server for layers module}
\usage{
layerServer(input, output, session, layers_selected, geom_blank_input,
  dataset, ggbase = NULL)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{layers_selected}{Reactive value of currently selected layer}

\item{geom_blank_input}{Need geom_blank values to check for inheritance}

\item{dataset}{Dataset}

\item{ggbase}{All your base are belong to us...}
}
\description{
Server for layers module
}
