% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layers.R
\name{layersServer}
\alias{layersServer}
\title{Server for all layers module}
\usage{
layersServer(id, dataset)
}
\arguments{
\item{id}{ID of layers module}

\item{dataset}{Dataset}
}
\description{
Server for all layers module
}
