% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layers.R
\name{layersServer}
\alias{layersServer}
\title{Module server for all layers}
\usage{
layersServer(input, output, session, dataset)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny user session}

\item{dataset}{Dataset}
}
\description{
Module server for all layers
}
