% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_layer.R
\name{layerUI}
\alias{layerUI}
\title{Module UI for individual layer}
\usage{
layerUI(id, server = FALSE, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Layer ID}

\item{server}{Render Shiny reactive elements if TRUE}

\item{session}{Shiny user session}
}
\value{
UI for layer
}
\description{
Module UI for individual layer
}
