% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prettyToggle}
\alias{prettyToggle}
\title{Wrapper for shinyWidgets::prettyToggle}
\usage{
prettyToggle(inputId, label_on, label_off, icon_on = NULL,
  icon_off = NULL, value = FALSE, status_on = "success",
  status_off = "danger", shape = c("square", "curve", "round"),
  outline = FALSE, fill = FALSE, thick = FALSE, plain = FALSE,
  bigger = FALSE, animation = NULL, inline = FALSE, width = NULL)
}
\value{
prettyToggle
}
\description{
We need to attach the header-icon class to the outermost div
}
