% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenity.viz.R
\docType{package}
\name{serenity.viz}
\alias{serenity.viz}
\alias{serenity.viz-package}
\title{Serenity Viz}
\usage{
serenity.viz(dataset = NULL)
}
\description{
Serenity Viz

Launch Serenity Viz in the default browser.
}
\examples{
\dontrun{
serenity.viz::serenity.viz()
}
}
