% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenity.viz.R
\docType{package}
\name{serenity.viz}
\alias{serenity.viz}
\title{Serenity Viz}
\usage{
serenity.viz(dataset = NULL)
}
\arguments{
\item{dataset}{Passed in dataset for visualization}
}
\description{
A platform-independent browser-based interface to data science tools for the classroom.
}
\examples{
\dontrun{
serenity.viz::serenity.viz()
}
}
