% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_serenity_viz.R
\name{serenityVizServer}
\alias{serenityVizServer}
\title{Server code for Serenity Viz module}
\usage{
serenityVizServer(input, output, session, dataset, trigger = NULL)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{dataset}{Passed in dataset for visualization}
}
\description{
Server code for Serenity Viz module
}
